+++ README.md
@@ -1,2 +1,2 @@
-# ros_rtt
+# micros_rtt
 A ros real-time toolkit based on orocos rtt.
