+++ package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>hpcl_rtt</name>
+  <name>micros_rtt</name>
   <version>0.0.4</version>
-  <description>The hpcl_rtt package</description>
+  <description>The micros_rtt package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
@@ -19,7 +19,7 @@
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <url type="git">https://github.com/sukha-cn/hpcl_rtt</url>
+  <url type="git">https://github.com/sukha-cn/micros_rtt</url>
 
 
   <!-- Author tags are optional, mutiple are allowed, one per tag -->
